requirements.txt
flask
Pillow
werkzeug
gunicorn